// frontend/lib/api/generated/instructionResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */
import type { InstructionResponseInstructionType } from "./instructionResponseInstructionType";
import type { InstructionResponseStatus } from "./instructionResponseStatus";

/**
 * 指令响应
 */
export interface InstructionResponse {
  /** 指令UUID */
  uuid: string;
  /** 会议UUID */
  conferenceId: string;
  /** 提交人UUID */
  submitterId: string;
  /** 提交人名称 */
  submitterName: string;
  /** 标题 */
  title: string;
  /** 指令类型 */
  instructionType: InstructionResponseInstructionType;
  /** 内容 */
  content: string;
  /** 状态 */
  status: InstructionResponseStatus;
  /** 批阅评语 */
  reviewComment?: string;
  /** 提交现实时间 */
  submitRealTime: string;
  /** 提交游戏时间 */
  submitGameTime: string;
  /** 最后批改人UUID */
  reviewedById?: string;
  /** 最后批改人名称 */
  reviewedByName?: string;
  /** 最后批改现实时间 */
  reviewedRealTime?: string;
  /** 最后批改游戏时间 */
  reviewedGameTime?: string;
}